[refactor] bot.pyをCog機構で機能別に分割する
1341行の単一ファイルを以下の構成に分割:
- bot_utils.py: 共通ユーティリティ（parse_date, create_embed等）
- cogs/health.py: ヘルスデータ照会コマンド（sleep, readiness等）
- cogs/report.py: レポート・分析コマンド（report, week, month等）
- cogs/settings_cog.py: 設定管理コマンド（goal, settings等）
- cogs/general.py: ヘルプ・自然言語対応
- cogs/scheduler.py: バックグラウンドタスク
